chore: move CallLabels to core
